add login styling, add error message when incorrect credentials

+++ src/components/LoginForm.vue
@@ -1,14 +1,21 @@
 <template>
   <div id="loginForm">
-    <c-toast-alert v-show="showErrorToast"></c-toast-alert>
+    <c-toast-alert v-show="showErrorToast" :content="errorMessage"></c-toast-alert>
+    <h1 class="form__title">Log in</h1>
     <form action="/post">
-      <input type="email" placeholder="Email address" v-model="username"/>
-      <input type="password" placeholder="Password" v-model="password">
-      <button type="submit" @click.prevent="login">Log in</button>
-      <label>
-        <input type="checkbox"/>
-        Keep me logged in
-      </label>
+      <div class="form-field">
+        <input type="email" placeholder="Email address" class="form__input" v-model="username"/>
+      </div>
+      <div class="form-field">
+        <input type="password" placeholder="Password" class="form__input" v-model="password"/>   
+      </div>
+      <div class="form-field">
+        <label class="form__label">
+          <input type="checkbox" class="form__checkbox" v-model="keepLoggedIn"/>
+          Keep me logged in
+        </label>
+      </div>   
+      <button class="form__login-btn" type="submit" @click.prevent="login">Login</button>
     </form>
     <c-loader v-if="loading"></c-loader>
   </div>
@@ -30,7 +37,9 @@ export default {
       loading: false,
       username: '',
       password: '',
-      showErrorToast: false
+      showErrorToast: false,
+      errorMessage: '',
+      keepLoggedIn: false
     }
   },
   methods: {
@@ -43,10 +52,13 @@ export default {
         this.$http.post('https://bench-api.applover.pl/api/v1/login', postableData).then( res => {
           this.loading = true;
           setTimeout(() => {
-            localStorage.setItem('login', true);
+            if (this.keepLoggedIn) {
+              localStorage.setItem('login', true);
+            }
             this.$router.push('/home')
           }, 1000)
         }, err => {
+            this.errorMessage = err.statusText
             this.showErrorToast = true;
             localStorage.setItem('login', false);
             this.$router.push('/login')
